services/miStudioScore/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
torch
transformers
datasets
scikit-learn
numpy